fix: cloud storage trpc schema and relocate storage helper
